# docs/api/openapi.json
{
  "components" : {
    "schemas" : {
...
      "AgentProfileResponse" : {
        "properties" : {
          "active" : {
            "type" : "boolean"
          },
          "agentId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "contextPolicy" : {
            "type" : "string"
          },
          "directives" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "domain" : {
            "type" : "string"
          },
          "limits" : {
            "type" : "string"
          },
          "model" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "outputFormat" : {
            "type" : "string"
          },
          "parentId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "resources" : {
            "items" : {
              "$ref" : "#/components/schemas/ResourceResponse"
            },
            "type" : "array"
          },
          "responsibilities" : {
            "type" : "string"
          },
          "role" : {
            "type" : "string"
          },
          "software" : {
            "items" : {
              "$ref" : "#/components/schemas/SoftwareRef"
            },
            "type" : "array"
          },
          "specialization" : {
            "type" : "string"
          },
          "systemPrompt" : {
            "type" : "string"
          },
          "version" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "Installed" : {
        "properties" : {
          "agentId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          },
          "outcome" : {
            "enum" : [ "CREATED", "EXISTING" ],
            "type" : "string"
          },
          "parentId" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "ProfileRequest" : {
        "properties" : {
          "autonomyLevel" : {
            "enum" : [ "GUIDED", "SUPERVISED", "DELEGATED", "FINAL_REVIEW" ],
            "type" : "string"
          },
          "projectType" : {
            "enum" : [ "WEB_APP", "BACKEND", "MOBILE", "DESKTOP", "AI_ML", "GAME", "THREE_D", "DATA", "AUTOMATION", "API", "FULL_STACK", "OTHER" ],
            "type" : "string"
          },
          "stack" : {
            "maxLength" : 2000,
            "minLength" : 0,
            "type" : "string"
          },
          "workspacePath" : {
            "maxLength" : 1000,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ProjectResponse" : {
        "properties" : {
          "autonomyLevel" : {
            "enum" : [ "GUIDED", "SUPERVISED", "DELEGATED", "FINAL_REVIEW" ],
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          },
          "planStatus" : {
            "enum" : [ "NONE", "DRAFT", "APPROVED" ],
            "type" : "string"
          },
          "projectType" : {
            "enum" : [ "WEB_APP", "BACKEND", "MOBILE", "DESKTOP", "AI_ML", "GAME", "THREE_D", "DATA", "AUTOMATION", "API", "FULL_STACK", "OTHER" ],
            "type" : "string"
          },
          "stack" : {
            "type" : "string"
          },
          "status" : {
            "enum" : [ "ACTIVE", "ARCHIVED" ],
            "type" : "string"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "workspacePath" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ResourceRequest" : {
        "properties" : {
          "configuration" : {
            "maxLength" : 4000,
            "minLength" : 0,
            "type" : "string"
          },
          "description" : {
            "maxLength" : 2000,
            "minLength" : 0,
            "type" : "string"
          },
          "key" : {
            "minLength" : 1,
            "pattern" : "^[a-z0-9][a-z0-9-]{1,63}$",
            "type" : "string"
          },
          "kind" : {
            "enum" : [ "SKILL", "KNOWLEDGE", "MCP", "TOOL", "FRAMEWORK", "TEMPLATE_PROVIDER" ],
            "type" : "string"
          },
          "name" : {
            "maxLength" : 120,
            "minLength" : 0,
            "type" : "string"
          },
          "searchUrl" : {
            "maxLength" : 500,
            "minLength" : 0,
            "pattern" : "^(https?://.*)?$",
            "type" : "string"
          },
          "sourceUrl" : {
            "maxLength" : 500,
            "minLength" : 0,
            "pattern" : "^(https?://.*)?$",
            "type" : "string"
          },
          "tags" : {
            "items" : {
              "maxLength" : 40,
              "minLength" : 0,
              "type" : "string"
            },
            "maxItems" : 20,
            "minItems" : 0,
            "type" : "array"
          }
        },
        "required" : [ "key", "kind", "name" ],
        "type" : "object"
      },
      "ResourceResponse" : {
        "properties" : {
          "configuration" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "key" : {
            "type" : "string"
          },
          "kind" : {
            "enum" : [ "SKILL", "KNOWLEDGE", "MCP", "TOOL", "FRAMEWORK", "TEMPLATE_PROVIDER" ],
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "searchUrl" : {
            "type" : "string"
          },
          "sourceUrl" : {
            "type" : "string"
          },
          "tags" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
...
      "SoftwareRef" : {
        "properties" : {
          "key" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearer" : {
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Every route needs `Authorization: Bearer <token>` (ADR-013). Every mutation of an existing resource needs `If-Match` with the ETag last read (ADR-009): absent is 428, stale is 412. Every error is `application/problem+json` with a stable `type` = `urn:ai-company-os:problem:<slug>` (ADR-007); branch on `type`, not on `title`.",
    "title" : "AI Company OS — control plane",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/agent-templates/{key}/install" : {
      "post" : {
        "operationId" : "install",
        "parameters" : [ {
          "in" : "path",
          "name" : "key",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/Installed"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "harness-controller" ]
      }
    },
...
    "/api/agents/{id}/profile" : {
      "get" : {
        "operationId" : "profile",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AgentProfileResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "harness-controller" ]
      },
      "put" : {
        "operationId" : "configure_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "in" : "header",
          "name" : "If-Match",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ProfileRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AgentProfileResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "harness-controller" ]
      }
    },
    "/api/agents/{id}/resources/{key}" : {
      "delete" : {
        "operationId" : "detach",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "in" : "path",
          "name" : "key",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "tags" : [ "harness-controller" ]
      },
      "put" : {
        "operationId" : "attach",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "in" : "path",
          "name" : "key",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "tags" : [ "harness-controller" ]
      }
    },
    "/api/agents/{id}/software/{key}" : {
      "delete" : {
        "operationId" : "disallow",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "in" : "path",
          "name" : "key",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "tags" : [ "harness-controller" ]
      },
      "put" : {
        "operationId" : "allow",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "in" : "path",
          "name" : "key",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "tags" : [ "harness-controller" ]
      }
    },
...
    "/api/projects/{id}/profile" : {
      "put" : {
        "operationId" : "configure",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "in" : "header",
          "name" : "If-Match",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ProfileRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjectResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "project-workspace-controller" ]
      }
    },
    "/api/projects/{id}/resources" : {
      "get" : {
        "operationId" : "projectResources",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/ResourceResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "harness-controller" ]
      }
    },
    "/api/projects/{id}/resources/{key}" : {
      "delete" : {
        "operationId" : "drop",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "in" : "path",
          "name" : "key",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "tags" : [ "harness-controller" ]
      },
      "put" : {
        "operationId" : "adopt",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "in" : "path",
          "name" : "key",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "tags" : [ "harness-controller" ]
      }
    },
...
    "/api/resources" : {
      "get" : {
        "operationId" : "search",
        "parameters" : [ {
          "in" : "query",
          "name" : "kind",
          "required" : false,
          "schema" : {
            "enum" : [ "SKILL", "KNOWLEDGE", "MCP", "TOOL", "FRAMEWORK", "TEMPLATE_PROVIDER" ],
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "q",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/ResourceResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "harness-controller" ]
      },
      "post" : {
        "operationId" : "create_1",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ResourceRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResourceResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "harness-controller" ]
      }
    },
...
  },
  "security" : [ {
    "bearer" : [ ]
  } ]
}